Update the tests for migrating around the handlers
