+++ src/sentry/workflow_engine/apps.py
@@ -6,6 +6,5 @@ class Config(AppConfig):
 
     def ready(self):
         # Import our base DataConditionHandlers for the workflow engine platform
-        import sentry.workflow_engine.action_handlers  # NOQA
-        import sentry.workflow_engine.condition_handlers  # NOQA
+        import sentry.workflow_engine.handlers  # NOQA
         from sentry.workflow_engine.endpoints import serializers  # NOQA
