+++ src/sentry/workflow_engine/handlers/__init__.py
@@ -1,0 +1,5 @@
+# Export any handlers we want to include into the registry
+__all__ = ["NotificationActionHandler", "GroupEventConditionHandler"]
+
+from .action import NotificationActionHandler
+from .condition import GroupEventConditionHandler
